refactor: 인수테스트 본문과 테스트 조각을 분리
